+++ components/Issues/IssueStateAuthorIcon.tsx
@@ -37,7 +37,7 @@ export const IssueStateAuthorIcon = ({ state, author }: { state: IssueState; aut
 
   return (
     <div className="relative w-max">
-      <Link href={`/user/${author}`} title={`User profile for ${author}`}>
+      <Link href={`/u/${author}`} title={`User profile for ${author}`}>
         <Avatar contributor={author} size="medium" />
       </Link>
       <div