+++ components/molecules/ContributorProfileHeader/contributor-profile-header.tsx
@@ -188,7 +188,7 @@ const ContributorProfileHeader = ({
           <div className="flex flex-col items-center gap-3 translate-y-24 md:translate-y-0 md:flex-row">
             <div className="flex flex-wrap items-center justify-center gap-2 mb-10 md:gap-6">
               <Button
-                onClick={() => handleCopyToClipboard(`${host}/user/${githubName}`)}
+                onClick={() => handleCopyToClipboard(`${host}/u/${githubName}`)}
                 className="my-auto gap-2 items-center shrink-0 place-self-end"
                 variant="primary"
               >