+++ pages/star-search/index.tsx
@@ -127,9 +127,14 @@ export async function getServerSideProps(context: GetServerSidePropsContext) {
   } = await supabase.auth.getSession();
 
   const userId = Number(session?.user.user_metadata.sub);
+  const searchParams = new URLSearchParams();
+
+  if (context.query.prompt) {
+    searchParams.set("prompt", context.query.prompt as string);
+  }
 
   const ogImageUrl = `${new URL(
-    "/assets/og-images/star-search-og-image.png",
+    `/og-images/star-search/?${searchParams}`,
     process.env.NEXT_PUBLIC_BASE_URL || "http://localhost:3000"
   )}`;
 
@@ -671,7 +676,7 @@ function SuggestionBoxes({
   suggestions: SuggesionTypes[];
 }) {
   return isHorizontal ? (
-    <Carousel className="w-fit max-w-[32rem] my-0 mx-auto px-auto" orientation="horizontal">
+    <Carousel className="w-fit max-w-[32rem] my-0 mx-auto px-auto md:ml-[1.63rem] lg:mx-auto" orientation="horizontal">
       <CarouselContent>
         {suggestions.map((suggestion, i) => (
           <CarouselItem key={i} className="items-stretch">